Upgraded to a Background Worker. It seems to be much more robust than system.threading. Also ran some debugging in the save() method where the primary keys and global ui variables were not being updated.
I think my biggest slowdown is constantly reading and writing to excel. I think I want to read once to a datatable and write back to xlsx when I'm done. This has the advantage of freeing the file when I'm done reading it; I don't need to keep a open, blank workbook in ram while I calculate. However, this works, and it will do for now until I can return. If I don't get time in the next few days, I'll return in March.